Added Maxine's House
Added house model and textures and set up the materials.

One important note, Virtual Texture support needs to be enabled in the project settings for the textures to show up properly. (Virtual Texture support doesn't change anything important, it just allows textures that are spread on multiple UV tiles to be imported properly)